refactor(db): extract authentication allowlist store
